requirements.txt
langchain==0.0.66
weaviate-client
openai
black
isort
Flask
transformers
gradio
python-dotenv
tiktoken
